♻️refactor: MoreListPage 레이아웃 및 소스 정리
