+++ src/app/more/MoreList.constants.ts
@@ -1,8 +1,5 @@
-// 임시 데이터
-export interface Dummy {
-  postId: number;
-  id: number;
-  name: string;
-  email: string;
-  body: string;
-}
+export const HEADER_TITLE: Record<string, string> = {
+  my_log: "내 최근 산책 기록",
+  my_post: "내가 작성한 산책로",
+  my_like: "내가 좋아하는 산책로",
+};
