+++ src/app/more/MoreList.types.ts
@@ -1,0 +1,8 @@
+// 임시 데이터
+export interface Dummy {
+  postId: number;
+  id: number;
+  name: string;
+  email: string;
+  body: string;
+}
